Release (#6)
* release candidate

This is v2.2.0-wsm and needs to be adapted to be compatible with Open Modelica

* Update FUNDING.yml

added PayPal Link

* Update FUNDING.yml

* delete redundant Icon and Diagram annotations

Great catch!

* Update ConstantConverterBoolean.mo

Remove redundant Icon() and Diagram() annotations

* Update package.mo

Fix missing </a> in hyperlink of documentation

* make type Icons invisible

Icon layer of a replaceable type should not be inherited by dependent classes

* fix output connector icons

output connectors with replaceable type OutputType should not inherit the letter "R" from the type component. Using IconMap at the right place this is now fixed.

* Update DataTransferRate.mo

corrected syntax error in revision annotation

* adapt to Open Modelica backend

Avoid array slicing in initial equation sections for DelayN. Remove "Press F1 for documentation" diagram displays.

* adapt examples

Adapt examples to Open Modelica and other tools, which do not allow for entering non-SI units of time like years or months. All models are now stricly in seconds, which are to be interpreted as "standard model unit of time".

* Update SoftwareReleaseProject.mo

add diagram display of unit of time interpretation

* fix missing diagram titles

Diagram titles and page definitions were lost in last commit and needed to be replaced

* Update SimpleProductionChain.mo

* preparing the release

Everything should now run smoothly in OMEdit.